move logic into LegacySubgraphService
